+++ src/main/java/org/snomed/snowstormlite/service/ecl/constraint/SSubExpressionConstraint.java
@@ -3,7 +3,8 @@ package org.snomed.snowstormlite.service.ecl.constraint;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.apache.lucene.search.BooleanClause;
 import org.apache.lucene.search.BooleanQuery;
-import org.apache.lucene.search.MatchAllDocsQuery;
+import org.apache.lucene.index.Term;
+import org.apache.lucene.search.TermQuery;
 import org.snomed.langauges.ecl.domain.expressionconstraint.SubExpressionConstraint;
 import org.snomed.langauges.ecl.domain.filter.ConceptFilterConstraint;
 import org.snomed.langauges.ecl.domain.filter.DescriptionFilterConstraint;
@@ -76,7 +77,8 @@ public class SSubExpressionConstraint extends SubExpressionConstraint implements
 
 	private BooleanQuery.Builder doAddQuery(BooleanQuery.Builder builder, ExpressionConstraintLanguageService eclService) throws IOException {
 		if (wildcard) {
-			builder.add(new MatchAllDocsQuery(), BooleanClause.Occur.MUST);
+			// Match active concepts only, in line with Snowstorm, where ECL is evaluated over active concepts.
+			builder.add(new TermQuery(new Term(FHIRConcept.FieldNames.ACTIVE, "1")), BooleanClause.Occur.MUST);
 		} else if (conceptId != null) {
 			addConstraint(conceptId, builder, eclService);
 		} else if (nestedExpressionConstraint != null) {