+++ src/test/java/org/snomed/snowstormlite/ECLTest.java
@@ -38,7 +38,8 @@ class ECLTest {
 
 	@Test
 	void testECLWildcard() throws IOException {
-		assertEquals(25, valueSetService.expand("http://snomed.info/sct?fhir_vs=ecl/*", null, EN_LANGUAGE_DIALECTS, false, 0, 20).getExpansion().getTotal());
+		// 25 concepts in the test release, one of them inactive (75521003); the wildcard matches active concepts only
+		assertEquals(24, valueSetService.expand("http://snomed.info/sct?fhir_vs=ecl/*", null, EN_LANGUAGE_DIALECTS, false, 0, 20).getExpansion().getTotal());
 	}
 
 	@Test